MtrlFile: Split ColorTable apart from ColorSet
